Latest version.  Minor fix in AbstractAccessor.java.

+++ deps/src/ucar/trunk/multiarray/AbstractAccessor.java
@@ -1,4 +1,4 @@
-// $Id: AbstractAccessor.java,v 1.1.1.4 2000-08-28 21:54:42 dglo Exp $
+// $Id: AbstractAccessor.java,v 1.2 2002-05-29 20:32:38 steve Exp $
 /*
  * Copyright 1997-2000 Unidata Program Center/University Corporation for
  * Atmospheric Research, P.O. Box 3000, Boulder, CO 80307,
@@ -34,8 +34,8 @@ import java.io.IOException;
  * 
  * @see Accessor
  *
- * @author $Author: dglo $
- * @version $Revision: 1.1.1.4 $ $Date: 2000-08-28 21:54:42 $
+ * @author $Author: steve $
+ * @version $Revision: 1.2 $ $Date: 2002-05-29 20:32:38 $
  */
 
 public abstract class
@@ -255,7 +255,7 @@ AbstractAccessor
 	setDouble(int [] index, double value)
 		throws IOException
 	{
-		set(index, new Float(value));
+		set(index, new Double(value));
 	}
 
  /* End MultiArray write access methods */
